Rename shapes folder
