+++ napari/layers/shapes/shape_list.py
@@ -1,5 +1,5 @@
 import numpy as np
-from .shapes import Shape, Rectangle, Ellipse, Line, Path, Polygon
+from .shape_models import Shape, Rectangle, Ellipse, Line, Path, Polygon
 from .shape_util import inside_triangles, triangles_intersect_box
 from .mesh import Mesh
