+++ napari/layers/shapes/shape_models/__init__.py
