+++ napari/layers/shapes/shape_models/ellipse.py
