+++ napari/layers/shapes/shape_models/line.py
