+++ napari/layers/shapes/shape_models/path.py
