+++ napari/layers/shapes/shape_models/polygon.py
