+++ napari/layers/shapes/shape_models/rectangle.py
