+++ napari/layers/shapes/shape_models/shape.py
